+++ code/src/cljs/sixsq/nuvla/ui/apps/views_versions.cljs
@@ -19,16 +19,6 @@
      (str/capitalize (@tr [:versions]))]))
 
 
-(defn show-versions [show-versions?]
-  (let [tr        (subscribe [::i18n-subs/tr])
-        label     (@tr (if @show-versions? [:hide-versions] [:show-versions]))
-        icon-name (if @show-versions? "caret down" "caret right")]
-    [:a {:style    {:cursor "pointer"}
-         :on-click #(reset! show-versions? (not @show-versions?))}
-     [ui/Icon {:name icon-name}]
-     label]))
-
-
 (defn version-comparison-modal
   [left right version-left version-right]
   (if (and left right)